extracted parser into Parser
